% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/address.R
\name{address}
\alias{address}
\title{Gets the internal memory address of a variable}
\usage{
address(var)
}
\arguments{
\item{var}{A variable}
}
\value{
A character string
}
\description{
Gets the internal memory address of a variable
}

