% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest_elements.R
\name{digest_list}
\alias{digest_list}
\title{Derive digests of argument values or their contents}
\usage{
digest_list(list, funs.unwrap = NULL)
}
\arguments{
\item{list}{Named list of elements to digest}

\item{funs.unwrap}{Named list of functions to be applied to elements (matched on names) in `args`. Optional.}
}
\value{
A data.table with columns 'name' and 'res' (digest values)
}
\description{
Get digests of argument values. Optionally, functions can be run
on some arguments before calculating digests. An example would be
reading contents of a file where the file path is an argument.
}
\keyword{internal}
