% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/list_primitives.R
\name{list_primitives}
\alias{list_primitives}
\title{Lists all primitives in R}
\usage{
list_primitives(names = FALSE)
}
\arguments{
\item{names}{If TRUE, names are returned, otherwise functions.}
}
\value{
A character vector or a list of functions.
}
\description{
Lists all primitives in R
}
\details{
This functions uses \code{methods:::.BasicFunsList}.
}

