% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/inspect.R
\name{named}
\alias{named}
\title{Gets the NAMED status of a variable}
\usage{
named(var)
}
\arguments{
\item{var}{A variable}
}
\value{
An integer (0, 1, or 2).
}
\description{
Gets the NAMED status of a variable
}

