% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/recyclables.R
\name{recyclables}
\alias{recyclables}
\title{Test reyclability of "base" functions}
\usage{
recyclables(sets = c("1arg", "2args"), mode = c("double", "integer",
  "logical"), debug = FALSE)
}
\arguments{
\item{sets}{Types of functions to test.}

\item{mode}{Data type on input arguments used for testing.}

\item{debug}{If TRUE, debug information is outputted, otherwise not.}
}
\value{
A \link{data.frame}.
}
\description{
Test reyclability of "base" functions
}

