% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_need_run.R
\name{check_need_run}
\alias{check_need_run}
\title{Check if function needs to be re-run based on argument changes}
\usage{
check_need_run(args, path.res, path.digest, funs.unwrap, force)
}
\arguments{
\item{args}{Named list of arguments}

\item{path.res}{Path to results file}

\item{path.digest}{Path to digest file}

\item{funs.unwrap}{Named list of functions to apply before digesting}

\item{force}{Logical, if TRUE always return that run is needed}
}
\value{
List with elements: run (logical), digest.new (data.table), digest.all (data.table or NULL)
}
\description{
Compares current arguments against stored digests to determine if
a function needs to be re-executed.
}
\keyword{internal}
