% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recycle-package.R
\docType{package}
\name{recycle-package}
\alias{recycle-package}
\title{recycle: Set Hook on Garbage Collection}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The 'recycle' function lets you call a function of your choice whenever garbacge collection is called, with a minimum delta in seconds to limit the frequency. This can be done in the current session or in another session in the background and might be used for instance for periodic notifications, logs, diagnostics or updates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/moodymudskipper/recycle}
  \item Report bugs at \url{https://github.com/moodymudskipper/recycle/issues}
}

}
\author{
\strong{Maintainer}: Antoine Fabri \email{antoine.fabri@gmail.com}

}
\keyword{internal}
