% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/recycle.R
\name{recycle}
\alias{recycle}
\title{Removes variable but recycles its value iff possible}
\usage{
recycle(var)
}
\arguments{
\item{var}{A variable}
}
\value{
The value of the recycled variable
}
\description{
Removes variable but recycles its value iff possible
}
\examples{
## Creates a variable
x <- rnorm(1e6); x[1] <- 1
print(ax <- address(x))
## Removes the variable, but recycle its values (iff possible)
y <- recycle(x)
print(ay <- address(y))
stopifnot(ay == ax)
}

