% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle.R, R/recycle.R
\name{new_cycle}
\alias{new_cycle}
\alias{new_cycle_bg}
\alias{recycle}
\alias{uncycle}
\alias{cycles}
\title{Set a hooks on garbage collection}
\usage{
new_cycle(f, ..., delta = 1)

new_cycle_bg(f, ..., delta = 1)

recycle(...)

uncycle()

cycles()
}
\arguments{
\item{f}{The hook, a function to call whenever garbage collection is triggered.
For \code{new_cycle_bg()} The default values of arguments are evaluated when the hook
is called and copied to the newly created session.}

\item{...}{For \code{recycle()}: hooks to be added or replaced, defined with \code{cycle()} or
\code{cycle_bg()}. Name them or they'll be names using \code{make.names()}.
Unused in \code{new_cycle()} and \code{new_cycle_bg()}.}

\item{delta}{Minimal number of seconds between calls to the hook.}
}
\value{
Returns \code{cycles()} invisibly, called for side effects.
}
\description{
\code{recycle()} sets the hooks using \code{new_cycle()} and \code{new_cycle_bg()}. \code{cycles()}
displays the existing hooks. \code{uncycle()} cleans everything up.
\code{cycle_process()} returns an actual process object for advanced use.
}
\details{
\itemize{
\item \code{new_cycle()} defines a hook on garbage collection
\item \code{new_cycle_bg()} is similar but the hook will be run in the background,
in a different session. The output is redirected to the current session,
so if necessary, divert, capture or suppress it in \code{f}
\item \code{recycle()} activates the hooks, if calls several times it will overwrite,
add, or remove hooks. Hooks are run in order when deltas match.
\item \code{cycle_process()} returns a process, of class "<r_process/process/R6>"
documented in the 'processx' package
}
}
