% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recycle.R
\name{recycle}
\alias{recycle}
\title{Set a hook on garbage collection}
\usage{
recycle(f, delta = 1)
}
\arguments{
\item{f}{An argument-less function to call whenever garbage collection is triggered}
}
\description{
Set a hook on garbage collection
}
